requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.douban.com/simple
--trusted-host pypi.douban.com
amqp==5.0.5; python_version >= '3.6'
ansible-base==2.10.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
ansible==3.2.0
asgiref==3.3.1; python_version >= '3.5'
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7; python_version >= '3.6'
cycler==0.10.0
dateutils==0.6.12
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-templated-mail==1.1.1
django==3.1.7; python_version >= '3.6'
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.4
djoser==2.1.0
elasticsearch-dsl==7.3.0
elasticsearch==7.12.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itypes==1.2.0
jinja2==2.11.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
kafka-python==2.0.2
kiwisolver==1.3.1; python_version >= '3.6'
kombu==5.0.2; python_version >= '3.6'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
matplotlib==3.4.1
mysqlclient==2.0.3
numpy==1.20.2; python_version >= '3.7'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
phonenumbers==8.12.20
pillow==8.1.2; python_version >= '3.6'
prompt-toolkit==3.0.18; python_full_version >= '3.6.1'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.0.1; python_version >= '3.6'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-ldap==3.3.1
python3-openid==3.2.0; python_version >= '3.0'
pytz==2021.1
# pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
redis==3.5.3
requests-oauthlib==1.3.0
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==4.0.0
social-auth-core==4.1.0; python_version >= '3.6'
soupsieve==2.2.1; python_version >= '3.0'
sqlparse==0.4.1; python_version >= '3.5'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_full_version < '4.0.0'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
